fix: checked arithmetic for numeric conversions and overflow protection
- Add safe_float_to_i64/safe_float_to_u32 for bounded CBOR numeric decoding
- Use saturating arithmetic for exp/nbf tolerance to prevent overflow
- Make CatTokenBuilder::build() fallible with geo-coordinate and DPoP validation
- Make with_clock_skew_tolerance/with_separate_tolerances return Result
- Change CatRenewal expadd/deadline from i64 to f64 to preserve CBOR semantics

+++ src/claims.rs
@@ -280,8 +280,8 @@ impl CatRenewalType {
 #[derive(Debug, Clone, PartialEq, Serialize)]
 pub struct CatRenewal {
     pub renewal_type: CatRenewalType,
-    pub expadd: Option<i64>,
-    pub deadline: Option<i64>,
+    pub expadd: Option<f64>,
+    pub deadline: Option<f64>,
     pub cookie_name: Option<String>,
     pub header_name: Option<String>,
     pub cookie_params: Option<Vec<String>>,
@@ -342,12 +342,12 @@ impl CatRenewal {
         }
     }
 
-    pub fn with_expadd(mut self, seconds: i64) -> Self {
+    pub fn with_expadd(mut self, seconds: f64) -> Self {
         self.expadd = Some(seconds);
         self
     }
 
-    pub fn with_deadline(mut self, timestamp: i64) -> Self {
+    pub fn with_deadline(mut self, timestamp: f64) -> Self {
         self.deadline = Some(timestamp);
         self
     }
